Move unit tests to their appropriate location
